chap 14

+++ engine/src/ecs/systems/tilemap_renderer_system.h
@@ -30,7 +30,8 @@ namespace fuse::ecs {
         FUSE_INLINE void
         draw_tile(asset_id tileset_id, float x, float y, int row, int col,
                   int size, SDL_RendererFlip flip) {
-            SDL_FRect dst_rect = {x * size, y * size, size, size};
+            SDL_FRect dst_rect = {x * size, y * size, (float) size,
+                                  (float) size};
             SDL_Rect src_rect = {row * size, col * size, size, size};
             auto &tileset = _assets->get<texture_asset>(tileset_id)->instance;
             SDL_RenderCopyExF(_renderer, tileset.data, &src_rect, &dst_rect, 0,
